Перенесли все константы в один файл
