+++ src/presenter/event.js
@@ -1,12 +1,8 @@
 import EventView from "../view/event.js";
 import EditFormView from "../view/edit-form.js";
+import {WorkMode, RenderPosition} from "../const.js";
 
-import {render, RenderPosition, replace, remove} from "../utils/render.js";
-
-const Mode = {
-  DEFAULT: `DEFAULT`,
-  EDITING: `EDITING`
-};
+import {render, replace, remove} from "../utils/render.js";
 
 
 export default class EventPresenter {
@@ -17,7 +13,7 @@ export default class EventPresenter {
 
     this._eventComponent = null;
     this._eventEditComponent = null;
-    this._mode = Mode.DEFAULT;
+    this._mode = WorkMode.DEFAULT;
 
     this._handleEditClick = this._handleEditClick.bind(this);
     this._handleCancelEditClick = this._handleCancelEditClick.bind(this);
@@ -47,11 +43,11 @@ export default class EventPresenter {
       return;
     }
 
-    if (this._mode === Mode.DEFAULT) {
+    if (this._mode === WorkMode.DEFAULT) {
       replace(this._eventComponent, prevEventComponent);
     }
 
-    if (this._mode === Mode.EDITING) {
+    if (this._mode === WorkMode.EDITING) {
       replace(this._eventEditComponent, prevEventEditComponent);
     }
 
@@ -67,7 +63,7 @@ export default class EventPresenter {
 
 
   resetView() {
-    if (this._mode !== Mode.DEFAULT) {
+    if (this._mode !== WorkMode.DEFAULT) {
       this._replaceFormToCard();
     }
   }
@@ -77,14 +73,14 @@ export default class EventPresenter {
     replace(this._eventEditComponent, this._eventComponent);
     document.addEventListener(`keydown`, this._escKeyDownHandler);
     this._changeMode();
-    this._mode = Mode.EDITING;
+    this._mode = WorkMode.EDITING;
   }
 
 
   _replaceFormToCard() {
     replace(this._eventComponent, this._eventEditComponent);
     document.removeEventListener(`keydown`, this._escKeyDownHandler);
-    this._mode = Mode.DEFAULT;
+    this._mode = WorkMode.DEFAULT;
   }
